@@ -32,41 +32,42 @@
   <properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
-		<maven-jar-plugin.version>3.2.0</maven-jar-plugin.version>
-		<maven-javadoc-plugin.version>2.10.4</maven-javadoc-plugin.version>
-		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
+		<maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
+		<maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
+		<maven-javadoc-plugin.version>3.5.0</maven-javadoc-plugin.version>
+		<maven-gpg-plugin.version>3.0.1</maven-gpg-plugin.version>
 		<nexus-staging-maven-plugin.version>1.6.3</nexus-staging-maven-plugin.version>
-		<maven-dependency-plugin.version>2.10</maven-dependency-plugin.version>
-		<java.version>1.8</java.version>
+		<maven-dependency-plugin.version>3.5.0</maven-dependency-plugin.version>
+		<java.version>17</java.version>
 		<junit.version>4.13.2</junit.version>
-		<slf4j-api.versin>1.7.25</slf4j-api.versin>
-		<log4j-web.version>2.11.2</log4j-web.version>
-		<springboot.version>2.4.1</springboot.version>
-		<openfeign.version>3.0.0</openfeign.version>
-		<spring.version>5.3.2</spring.version>
-		<aspectj.version>1.8.13</aspectj.version>
+		<slf4j-api.versin>2.0.7</slf4j-api.versin>
+		<springboot.version>3.0.5</springboot.version>
+		<openfeign.version>4.0.2</openfeign.version>
+		<spring.version>6.0.7</spring.version>
+		<aspectj.version>1.9.19</aspectj.version>
 		<concurrentlinkedhashmap-lru.version>1.4.2</concurrentlinkedhashmap-lru.version>
-		<shiro.version>1.8.0</shiro.version>
-		<httpcomponents.version>4.5.6</httpcomponents.version>
-		<poi.version>5.0.0</poi.version>
-		<poi-ooxml.version>5.0.0</poi-ooxml.version>
-		<validation.version>1.1.0.Final</validation.version>
-		<mybatis.version>2.1.4</mybatis.version>
-		<pagehelper.version>1.3.0</pagehelper.version>
-		<druid.version>1.2.6</druid.version>
-		<webjars.version>0.42</webjars.version>
+		<shiro.version>2.0.0-alpha-1</shiro.version>
+		<httpcomponents.version>5.2.1</httpcomponents.version>
+		<http4components.version>4.5.14</http4components.version>
+		<poi.version>5.2.3</poi.version>
+		<poi-ooxml.version>5.2.3</poi-ooxml.version>
+		<commons-fileupload.version>1.5</commons-fileupload.version>
+		<commons-io.version>2.11.0</commons-io.version>
+		<mybatis.version>3.0.1</mybatis.version>
+		<pagehelper.version>1.4.6</pagehelper.version>
+		<druid.version>1.2.16</druid.version>
+		<webjars.version>0.46</webjars.version>
 		<quartz.version>2.3.2</quartz.version>
 		<gson.version>2.8.5</gson.version>
-		<activiti.version>5.22.0</activiti.version>
-		<batik.version>1.7</batik.version>
-		<junx.version>1.0.9</junx.version>
-		<kaptcha.version>0.0.9</kaptcha.version>
-		<fastjson.version>1.2.70</fastjson.version>
-		<ST.version>4.1</ST.version>
-		<thumbnailator.version>0.4.11</thumbnailator.version>
-		<imageio.version>3.5</imageio.version>
-		<QLExpress.version>3.2.4</QLExpress.version>
+		<junx.version>1.0.10</junx.version>
+		<kaptcha.version>2.3.0</kaptcha.version>
+		<fastjson.version>2.0.28</fastjson.version>
+		<ST.version>4.3.4</ST.version>
+		<thumbnailator.version>0.4.19</thumbnailator.version>
+		<imageio.version>3.9.4</imageio.version>
+		<QLExpress.version>3.3.1</QLExpress.version>
+		<freemarker.version>2.3.32</freemarker.version>
+		<apacheant.version>1.10.13</apacheant.version>
 	</properties>
   <dependencies>
 	<dependency>
