@@ -1,12 +1,12 @@
 package io.github.junxworks.ep.core.security.access;
 
-import javax.servlet.DispatcherType;
-
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
+
+import jakarta.servlet.DispatcherType;
 
 @Configuration
 @EnableConfigurationProperties({ AcConfig.class })
