@@ -16,7 +16,7 @@ public interface FileMapper extends BaseMapper{
 	public List<EpSFile> getAll();
 
 	@Select("select * from ep_s_file where id=#{id}")
-	public EpSFile findById(String id);
+	public EpSFile findById(Long id);
 
 	@Select("select * from ep_s_file where file_group=#{fileGroup}")
 	public List<EpSFile> findByGroup(String fileGroup);
@@ -25,11 +25,11 @@ public interface FileMapper extends BaseMapper{
 	public List<EpSFile> findByOrg(String orgNo);
 
 	@Select("delete from ep_s_file where id=#{id}")
-	public void deleteById(String id);
+	public void deleteById(Long id);
 
 	@Select("delete from ep_s_file where file_group=#{fileGroup}")
 	public void deleteByGroup(String fileGroup);
 
 	@Select("select * from ep_s_file_thumb where file_id=#{fileId} and width=#{width} and height=#{height} order by id desc")
-	public List<EpSFileThumb> queryThumbList(@Param("fileId")String fileId,@Param("width")int width,@Param("height")int height);
+	public List<EpSFileThumb> queryThumbList(@Param("fileId")Long fileId,@Param("width")int width,@Param("height")int height);
 }
